+++ grpc-server-runtime/src/main/java/io/micronaut/grpc/server/GrpcServerBuilder.java
@@ -25,7 +25,7 @@ import io.micronaut.context.annotation.Factory;
 import io.micronaut.core.annotation.Nullable;
 import io.micronaut.core.order.OrderUtil;
 import io.micronaut.core.util.CollectionUtils;
-import io.micronaut.grpc.health.HealthStatusManagerContainer;
+import io.micronaut.grpc.server.health.HealthStatusManagerContainer;
 import jakarta.inject.Inject;
 import jakarta.inject.Singleton;
 import org.slf4j.Logger;
