+++ grpc-server-runtime/src/main/java/io/micronaut/grpc/server/GrpcServerConfiguration.java
@@ -54,6 +54,8 @@ public class GrpcServerConfiguration {
     public static final String PORT = PREFIX + ".port";
     public static final String HOST = PREFIX + ".host";
     public static final String ENABLED = PREFIX + ".enabled";
+    @Deprecated
+    public static final String HEALTH_ENABLED = PREFIX + ".health.enabled";
     public static final int DEFAULT_PORT = 50051;
     public static final Duration DEFAULT_AWAIT_TERMINATION = Duration.ofSeconds(30);
